feat: add region selection, and implement collaborative filtering
