+++ src/components/Layout/Layout.scss
@@ -10,3 +10,7 @@
   margin: auto;
   padding: 2rem;
 }
+
+.header__title {
+  margin-top: 0;
+}
