+++ src/components/Layout/Layout.tsx
@@ -4,6 +4,9 @@ import './Layout.scss';
 export const Layout = () => {
   return (
     <div className="container">
+      <header className="header">
+        <h1 className="header__title">Countries Wiki</h1>
+      </header>
       <main>
         <Outlet />
       </main>
