+++ src/components/Main/Main.scss
@@ -1,0 +1,11 @@
+.navigation {
+  display: flex;
+  gap: 1rem;
+  padding-bottom: 1rem;
+}
+
+@media screen and (max-width: 768px) {
+  .navigation {
+    flex-direction: column;
+  }
+}
