+++ src/components/Main/Main.tsx
@@ -1,14 +1,16 @@
 import { useEffect } from 'react';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { CountryList } from './CountryList/CountryList';
 import { Search } from './Search/Search';
+import { SelectRegion } from './SelectRegion/SelectRegion';
 import { useGetCountriesQuery } from '@store/api';
-import { setCountries } from '@store/countriesSlice';
+import { setCountries, selectCountriesTime } from '@store/countriesSlice';
 import './Main.scss';
 
 export const Main = () => {
   const { data } = useGetCountriesQuery({});
   const dispatch = useDispatch();
+  const filteredCountries = useSelector(selectCountriesTime);
 
   useEffect(() => {
     if (data) {
@@ -18,8 +20,11 @@ export const Main = () => {
 
   return (
     <article>
-      <Search />
-      {data && <CountryList />}
+      <section className="navigation">
+        <Search />
+        <SelectRegion />
+      </section>
+      {filteredCountries && <CountryList />}
     </article>
   );
 };
