+++ src/components/Main/Search/Search.scss
@@ -1,7 +1,8 @@
 .search {
   display: flex;
   justify-content: center;
-  align-items: center;
+  align-items: flex-end;
+  flex: 1;
 }
 
 .search__input {
