+++ src/components/Main/SelectRegion/SelectRegion.scss
@@ -1,0 +1,35 @@
+.select-region {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.select-region__title {
+  font-size: 16px;
+  font-weight: bold;
+  margin: auto;
+}
+
+.select-region__select {
+  height: 40px;
+  padding: 8px;
+  font-size: 14px;
+  border-radius: 4px;
+  border: 1px solid #ccc;
+  outline: none;
+}
+
+.select-region__select:focus {
+  border-color: #007bff;
+  box-shadow: 0 0 5px rgba(0, 123, 255, 0.5);
+}
+
+.select-region__option {
+  font-size: 14px;
+}
+
+@media screen and (max-width: 768px) {
+  .navigation {
+    flex-direction: column;
+  }
+}
