+++ src/store/api.ts
@@ -1,5 +1,5 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-import { Country } from '@/src/types/types';
+import { Country } from '@/src/types/countryTypesAPI';
 
 export const countriesApi = createApi({
   reducerPath: 'countriesApi',
