+++ src/types/types.ts
@@ -1,115 +1,3 @@
-import { Status, Region, StartOfWeek } from './enums';
+import { Region } from '@/src/types/enums';
 
-export interface Country {
-  name: Name;
-  tld?: string[];
-  cca2: string;
-  ccn3?: string;
-  cca3: string;
-  independent?: boolean;
-  status: Status;
-  unMember: boolean;
-  currencies?: { [key: string]: Currency };
-  idd: Idd;
-  capital?: string[];
-  altSpellings: string[];
-  region: Region;
-  languages?: { [key: string]: string };
-  translations: { [key: string]: Translation };
-  latlng: number[];
-  landlocked: boolean;
-  area: number;
-  demonyms?: Demonyms;
-  flag: string;
-  maps: Maps;
-  population: number;
-  car: Car;
-  timezones: string[];
-  continents: Continent[];
-  flags: Flags;
-  coatOfArms: CoatOfArms;
-  startOfWeek: StartOfWeek;
-  capitalInfo: CapitalInfo;
-  cioc?: string;
-  subregion?: string;
-  fifa?: string;
-  borders?: string[];
-  gini?: { [key: string]: number };
-  postalCode?: PostalCode;
-}
-
-export interface CapitalInfo {
-  latlng?: number[];
-}
-
-export interface Car {
-  signs?: string[];
-  side: Side;
-}
-
-export enum Side {
-  Left = 'left',
-  Right = 'right',
-}
-
-export interface CoatOfArms {
-  png?: string;
-  svg?: string;
-}
-
-export enum Continent {
-  Africa = 'Africa',
-  Antarctica = 'Antarctica',
-  Asia = 'Asia',
-  Europe = 'Europe',
-  NorthAmerica = 'North America',
-  Oceania = 'Oceania',
-  SouthAmerica = 'South America',
-}
-
-export interface Currency {
-  name: string;
-  symbol: string;
-}
-
-export interface Demonyms {
-  eng: Eng;
-  fra?: Eng;
-}
-
-export interface Eng {
-  f: string;
-  m: string;
-}
-
-export interface Flags {
-  png: string;
-  svg: string;
-  alt?: string;
-}
-
-export interface Idd {
-  root?: string;
-  suffixes?: string[];
-}
-
-export interface Maps {
-  googleMaps: string;
-  openStreetMaps: string;
-}
-
-export interface Name {
-  common: string;
-  official: string;
-  nativeName?: { [key: string]: Translation };
-}
-
-export interface Translation {
-  official: string;
-  common: string;
-}
-
-export interface PostalCode {
-  format: string;
-  regex?: string;
-}
+export type RegionSelect = 'All' | Region;
